# CUDA: Add launch bounds for Pascal, small q4_K, q5_K refactor