Note: this PR is based on https://github.com/ggerganov/llama.cpp/pull/2590 so the diff shown on Github includes the changes from both PRs.

This PR adds `__launch_bounds__` to several mmq kernels for Pascal which informs the compiler of the desired number of kernels to fit on a single streaming multiprocessor so it can limit register usage accordingly. This limit serves to fit more kernels onto a single multiprocessor (at possibly larger tile sizes). There is also a small refactor of q4_K and q5_K that makes q4_K a little bit faster:

| GPU      | Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------|--------|---------|
| RTX 3090 | 7b q4_k_s | pp   |       1974 |   1994 |    1.01 |
| P40      | 7b q4_1   | pp   |        812 |    863 |    1.06 |
| P40      | 7b q3_k_s | pp   |        468 |    536 |    1.15 |
| P40      | 7b q4_k_s | pp   |        706 |    813 |    1.15 |
| P40      | 7b q6_k   | pp   |        626 |    752 |    1.20 |

mmq performance is now consistently better than cuBLAS for all quantization types (if you don't count f16). So I'll look into making it the default. What's the desired way to do this? Something like a `--blas` command line option that switches to a BLAS library if it's been compiled with (eventually not just cuBLAS)?