@LostRuins 2023-08-15
>  What's the desired way to do this? Something like a --blas command line option that switches to a BLAS library if it's been compiled with (eventually not just cuBLAS)?

You can probably just change main.cpp to deprecate the `-mmq` or `--mul-mat-q` parameter, and then add a `-nommq` `--no-mul-mat-q` parameter, making it true by default unless the flag is set. Though you might want to indicate it only applies to CUDA since all other backends that don't use mmq would still want to use with respective openblas/clblast implementations.